% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisao.R
\name{divisao}
\alias{divisao}
\title{Função para dividir.}
\usage{
divisao(a, b)
}
\arguments{
\item{a, }{numero que será dividido}

\item{b, }{outro número que irar dividir}
}
\description{
Esta função serve para dividir um numero por outro
}
