% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacao.R
\name{multiplicacao}
\alias{multiplicacao}
\title{Função para multiplicar.}
\usage{
multiplicacao(a, b)
}
\arguments{
\item{a, }{numero a multiplicar}

\item{b, }{outro número a multiplicar}
}
\description{
Esta função serve para multiplicar dois números
}
\examples{

multiplicacao(4,2)
resposta 8

}
